+++ src/commands/registerCommands.ts
@@ -45,9 +45,9 @@ import { redeployDeployment } from './deployments/redeployDeployment';
 import { viewCommitInGitHub } from './deployments/viewCommitInGitHub';
 import { viewDeploymentLogs } from './deployments/viewDeploymentLogs';
 import { editAppSetting } from './editAppSetting';
-import { EventGridCodeLensProvider } from './executeFunction/EventGridCodeLensProvider';
+import { EventGridCodeLensProvider } from './executeFunction/eventGrid/EventGridCodeLensProvider';
+import { sendEventGridRequest } from './executeFunction/eventGrid/sendEventGridRequest';
 import { executeFunction } from './executeFunction/executeFunction';
-import { sendEventGridRequest } from './executeFunction/sendEventGridRequest';
 import { initProjectForVSCode } from './initProjectForVSCode/initProjectForVSCode';
 import { startStreamingLogs } from './logstream/startStreamingLogs';
 import { stopStreamingLogs } from './logstream/stopStreamingLogs';
